llmproxy: use static secret for health checks
