+++ enterprise/cmd/llm-proxy/internal/httpapi/BUILD.bazel
@@ -4,8 +4,8 @@ go_library(
     name = "httpapi",
     srcs = [
         "anthropic.go",
+        "diagnostics.go",
         "handler.go",
-        "healthz.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/llm-proxy/internal/httpapi",
     visibility = ["//enterprise/cmd/llm-proxy:__subpackages__"],
