--- enterprise/cmd/llm-proxy/internal/httpapi/healthz.go
@@ -1,31 +1,0 @@
-package httpapi
-
-import (
-	"context"
-
-	"github.com/sourcegraph/sourcegraph/internal/redispool"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
-)
-
-func healthz(ctx context.Context) error {
-	// Check redis health
-	rpool, ok := redispool.Cache.Pool()
-	if !ok {
-		return errors.New("redis: not available")
-	}
-	rconn, err := rpool.GetContext(ctx)
-	if err != nil {
-		return errors.Wrap(err, "redis: failed to get conn")
-	}
-	defer rconn.Close()
-
-	data, err := rconn.Do("PING")
-	if err != nil {
-		return errors.Wrap(err, "redis: failed to ping")
-	}
-	if data != "PONG" {
-		return errors.New("redis: failed to ping: no pong received")
-	}
-
-	return nil
-}
