+++ enterprise/cmd/llm-proxy/shared/config.go
@@ -16,6 +16,8 @@ type Config struct {
 
 	Address string
 
+	DiagnosticsSecret string
+
 	Dotcom struct {
 		URL             string
 		AccessToken     string
@@ -47,6 +49,9 @@ type TraceConfig struct {
 func (c *Config) Load() {
 	c.InsecureDev = env.InsecureDev
 	c.Address = c.Get("LLM_PROXY_ADDR", ":9992", "Address to serve LLM proxy on.")
+	c.DiagnosticsSecret = c.GetOptional("LLM_PROXY_DIAGNOSTICS_SECRET", "Secret for accessing diagnostics - "+
+		"should be used as 'Authorization: Bearer $secret' header when accessing diagnostics endpoints.")
+
 	c.Dotcom.AccessToken = c.Get("LLM_PROXY_DOTCOM_ACCESS_TOKEN", "", "The Sourcegraph.com access token to be used.")
 	c.Dotcom.URL = c.Get("LLM_PROXY_DOTCOM_API_URL", "https://sourcegraph.com/.api/graphql", "Custom override for the dotcom API endpoint")
 	c.Dotcom.DevLicensesOnly = c.GetBool("LLM_PROXY_DOTCOM_DEV_LICENSES_ONLY", "false", "Only allow tokens associated with active dev licenses to be used.")
