+++ enterprise/cmd/llm-proxy/shared/main.go
@@ -77,8 +77,11 @@ func Main(ctx context.Context, obctx *observation.Context, ready service.ReadyFu
 		Next:        handler,
 	}
 
+	// Diagnostic layers should come before auth/rate-limiting
+	handler = httpapi.NewDiagnosticsHandler(obctx.Logger, handler, config.DiagnosticsSecret)
+
 	// Instrumentation layers
-	handler = httpLogger(obctx.Logger.Scoped("httpAPI", ""), handler)
+	handler = requestLogger(obctx.Logger.Scoped("requests", "HTTP requests"), handler)
 	handler = instrumentation.HTTPMiddleware("llm-proxy", handler)
 
 	// Collect request client for downstream handlers. Outside of dev, we always set up
@@ -161,7 +164,7 @@ func rateLimit(logger log.Logger, eventLogger events.Logger, cache limiter.Redis
 	})
 }
 
-func httpLogger(logger log.Logger, next http.Handler) http.Handler {
+func requestLogger(logger log.Logger, next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		rc := requestclient.FromContext(r.Context())
 		logger.Debug("Request",
